[IMP] mrp_workorder_grouping_by_material: post_inventory on workorders
