+++ mrp_workorder_grouping_by_material/models/__init__.py
@@ -2,4 +2,6 @@ from . import mrp_workorder
 from . import mrp_workcenter
 from . import mrp_bom
 from . import mrp_workorder_nest
+from . import mrp_workorder_nest_line
+from . import mrp_production
 from . import stock_production_lot
