+++ mrp_workorder_grouping_by_material/models/mrp_workorder_nest.py
@@ -14,6 +14,7 @@ NEST_STATE = [
     ("progress", "In Progress"),
     ("blocked", "Blocked"),
     ("done", "Done"),
+    ("cancel", "Cancelled"),
 ]
 
 
@@ -133,10 +134,10 @@ class MrpWorkorderNest(models.Model):
             result.append((record.id, name))
         return result
 
-    @api.onchange("lot_id")
-    def onchange_lot_id(self):
-        for line in self.nested_line_ids:
-            line.lot_id = self.lot_id.id
+    # @api.onchange("lot_id")
+    # def onchange_lot_id(self):
+    #     for line in self.nested_line_ids:
+    #         line.lot_id = self.lot_id.id
 
     @api.model
     def create(self, vals):
@@ -236,8 +237,9 @@ class MrpWorkorderNest(models.Model):
             nest.line_is_produced = is_produced
 
     def get_worksheets(self):
-        self.ensure_one()
-        worksheets = self.mapped("nested_line_ids.workorder_id.worksheet")
+        worksheets = self.mapped("nested_line_ids").get_worksheet()
+        if not worksheets:
+            return
         merger = PdfFileMerger(strict=False)
         for worksheet in worksheets:
             temp = tempfile.NamedTemporaryFile(suffix=".pdf")
@@ -255,8 +257,11 @@ class MrpWorkorderNest(models.Model):
 
     def show_worksheets(self):
         self.ensure_one()
+        binary = self.get_worksheets()
+        if not binary:
+            return
         wizard = self.env["binary.container"].create({
-            "binary_field": self.get_worksheets(),
+            "binary_field": binary,
         })
         action = self.env.ref("pdf_previewer.binary_container_action")
         action_dict = action and action.read()[0] or {}
@@ -265,312 +270,3 @@ class MrpWorkorderNest(models.Model):
             "res_id": wizard.id,
         })
         return action_dict
-
-
-class MrpWorkorderNestLine(models.Model):
-    _name = "mrp.workorder.nest.line"
-
-    def _get_selection_workorder_state(self):
-        return self.env["mrp.workorder"].fields_get(
-            allfields=["state"])["state"]["selection"]
-
-    nest_id = fields.Many2one(
-        comodel_name="mrp.workorder.nest",
-        required=True)
-    related_qty_producing = fields.Float(
-        related="workorder_id.qty_producing")
-    related_finished_lot_id = fields.Many2one(
-        comodel_name="stock.production.lot",
-        related="workorder_id.finished_lot_id")
-    qty_producing = fields.Float(
-        string="Quantity Producing", default=1.0,
-        digits="Product Unit of Measure", copy=True)
-    finished_lot_id = fields.Many2one(
-        comodel_name="stock.production.lot", string="Lot/Serial Number",
-        domain="[('product_id', '=', product_id)]")
-    lot_id = fields.Many2one(
-        comodel_name="stock.production.lot",
-        domain="[('product_id', '=', nest_id.main_product_id.id)]")
-    workorder_id = fields.Many2one(
-        comodel_name="mrp.workorder")
-    qty_produced = fields.Float(
-        string="Quantity Produced",
-        related="workorder_id.qty_produced")
-    qty_nested = fields.Float(
-        string="Nested Quantity",
-        related="workorder_id.qty_nested")
-    name = fields.Char(related="workorder_id.name", string="Name")
-    workcenter_id = fields.Many2one(
-        comodel_name="mrp.workcenter",
-        related="workorder_id.workcenter_id",
-        string="Workcenter",
-        readonly=True)
-    date_planned_start = fields.Datetime(
-        related="workorder_id.date_planned_start",
-        string="Scheduled Date Start",
-        readonly=True)
-    production_id = fields.Many2one(
-        comodel_name="mrp.production",
-        related="workorder_id.production_id",
-        string="Production",
-        readonly=1)
-    product_id = fields.Many2one(
-        comodel_name="product.product",
-        related="workorder_id.product_id",
-        string="Product",
-        readonly=1)
-    product_tracking = fields.Selection(related="product_id.tracking")
-    qty_production = fields.Float(
-        string="Original Production Quantity",
-        related="workorder_id.qty_production",
-        readonly=1)
-    qty_remaining = fields.Float(
-        related="workorder_id.qty_remaining",
-        string="Quantity To Be Produced",
-        digits="Product Unit of Measure")
-    product_uom_id = fields.Many2one(
-        comodel_name="uom.uom",
-        related="workorder_id.product_uom_id",
-        string="Unit of Measure",
-        readonly=1)
-    state = fields.Selection(
-        string="State",
-        related="workorder_id.state",
-        selection=_get_selection_workorder_state,
-        readonly=True)
-    working_state = fields.Selection(
-        string="Workcenter Status",
-        related="workorder_id.working_state",
-        readonly=True)
-    is_produced = fields.Boolean(related="workorder_id.is_produced")
-    is_user_working = fields.Boolean(related="workorder_id.is_user_working")
-    production_state = fields.Selection(
-        related="workorder_id.production_state", readonly=1)
-    company_id = fields.Many2one(related="workorder_id.company_id")
-    ok_line = fields.Boolean(String="Ok")
-    possible_workorder_ids = fields.Many2many(
-        comodel_name="mrp.workorder",
-        compute="_compute_possible_workorder_ids",
-        string="Possible Workorders")
-    worksheet = fields.Binary(
-        string="Worksheet",
-        related="workorder_id.worksheet",
-        readonly=True)
-
-    @api.depends("nest_id", "nest_id.main_product_id",
-                 "nest_id.workcenter_id", "nest_id.workorder_ids")
-    def _compute_possible_workorder_ids(self):
-        workorder_obj = self.env["mrp.workorder"]
-        for nest_line in self:
-            nest = nest_line.nest_id
-            assigned_wo = nest.workorder_ids.ids
-            workorders = workorder_obj.search(
-                [("main_product_id", "=", nest.main_product_id.id),
-                 ("workcenter_id", "=", nest.workcenter_id.id),
-                 ("id", "not in", assigned_wo),
-                 ("state", "not in", ["done", "cancel"])])
-            nest_line.possible_workorder_ids = [(6, 0, workorders.ids or [])]
-
-    def nest_line_form_view(self, res_id):
-        view_ref = self.env["ir.model.data"].get_object_reference(
-            "mrp_workorder_grouping_by_material",
-            "mrp_workorder_nest_line_form")
-        view_id = view_ref and view_ref[1] or False,
-        return {
-            "name": "Workorder nest line",
-            "domain": [],
-            "res_model": "mrp.workorder.nest.line",
-            "res_id": res_id,
-            "type": "ir.actions.act_window",
-            "view_mode": "form",
-            "view_type": "form",
-            "view_id": view_id,
-            "context": {},
-            "target": "new",
-        }
-
-    def button_get_previous_line(self):
-        self.ensure_one()
-        nest_lines = self.nest_id.nested_line_ids
-        line_index = -1
-        for index, line in enumerate(nest_lines):
-            if self.id == line.id:
-                line_index = index - 1
-        res_id = nest_lines[line_index].id if line_index >= 0 else self.id
-        return self.nest_line_form_view(res_id)
-
-    def button_get_next_line(self):
-        self.ensure_one()
-        nest_lines = self.nest_id.nested_line_ids
-        line_index = -1
-        for index, line in enumerate(nest_lines):
-            if self.id == line.id:
-                line_index = index + 1
-        if -1 < line_index < len(nest_lines):
-            res_id = nest_lines[line_index].id
-        else:
-            res_id = self.id
-        return self.nest_line_form_view(res_id)
-
-    def get_worksheet(self):
-        self.ensure_one()
-        return self.workorder_id.worksheet
-
-    def show_worksheet(self):
-        self.ensure_one()
-        wizard = self.env["binary.container"].create({
-            "binary_field": self.get_worksheet(),
-        })
-        action = self.env.ref("pdf_previewer.binary_container_action")
-        action_dict = action and action.read()[0] or {}
-        action_dict.update({
-            "name": _("Worksheet"),
-            "res_id": wizard.id,
-        })
-        return action_dict
-
-    def _create_assign_lot(self, code, product_id):
-        if code:
-            lot_obj = self.env['stock.production.lot']
-            lot_id = False
-            if product_id.tracking == 'serial':
-                pass
-            if product_id.tracking == 'lot':
-                lot_id = lot_obj.search([
-                    ('name', '=', code),
-                    ('company_id', '=', self.env.company.id),
-                    ('product_id', '=', product_id.id)], limit=1).id
-            if product_id.tracking == 'none':
-                return False
-            if not lot_id:
-                return lot_obj.create({
-                    'name': code,
-                    'product_id': product_id.id,
-                    'company_id': self.env.company.id,
-                }).id
-            return lot_id
-
-    @api.model
-    def create(self, vals):
-        nest_id = vals.get('nest_id')
-        workorder_id = vals.get('workorder_id')
-        if nest_id and workorder_id:
-            nest = self.env['mrp.workorder.nest'].browse(nest_id)
-            code = "{}/{}".format(nest.name or "", nest.code or "")
-            product_id = self.env['mrp.workorder'].browse(
-                workorder_id).product_id
-            vals['finished_lot_id'] = self._create_assign_lot(code, product_id)
-        return super().create(vals)
-
-    def update_workorder_lines(self, line_values):
-        for values in line_values['to_create']:
-            values.pop('raw_workorder_id')
-            values.update({'workorder_id': self.workorder_id.id})
-            self.env['mrp.workorder.line'].create(values)
-        for line in line_values['to_delete']:
-            if line in self.raw_workorder_line_ids:
-                line.unlink()
-            else:
-                self.finished_workorder_line_ids -= line
-        for line, vals in line_values['to_update'].items():
-            line.write(vals)
-
-    def _write_lot_producing_qty(self):
-        for n_line in self:
-            if n_line.workorder_id.state not in ("done", "cancel"):
-                res = {
-                    "qty_producing": n_line.qty_producing,
-                }
-                line_values = n_line.workorder_id._update_workorder_lines()
-                n_line.update_workorder_lines(line_values)
-                if n_line.finished_lot_id:
-                    res.update({
-                        'finished_lot_id': n_line.finished_lot_id.id,
-                    })
-                if n_line.lot_id:
-                    workorder_lines = n_line.workorder_id.raw_workorder_line_ids
-                    move_line = workorder_lines.filtered(
-                        lambda x: x.product_id == n_line.nest_id.main_product_id)
-                    move_line.lot_id = n_line.lot_id
-                if res:
-                    n_line.workorder_id.write(res)
-
-    def button_finish(self):
-        for nl in self.filtered(
-                lambda n: n.is_produced and n.state == "progress"):
-            nl.workorder_id.with_context(from_nest=True).button_finish()
-
-    def _check_final_product_lot(self):
-        for wo in self.mapped("workorder_id"):
-            if not wo._check_final_product_lots():
-                UserError(_(
-                    '{}: You should provide a lot/serial number for the '
-                    'final product.').format(wo.name))
-
-    def button_start(self):
-        for nl in self.filtered(lambda n: n.state != "blocked"):
-            nl.workorder_id.with_context(from_nest=True).button_start()
-
-    def record_production(self):
-        for nl in self:
-            is_user_working = nl.workorder_id.is_user_working
-            is_produced = nl.workorder_id.is_produced
-            if is_user_working and nl.state == "progress" and not \
-                    is_produced:
-                nl._write_lot_producing_qty()
-                wo = nl.workorder_id
-                try:
-                    if wo.current_quality_check_id.quality_state == "none":
-                        wo.current_quality_check_id.do_pass()
-                    wo.with_context(from_nest=True).record_production()
-                except UserError as e:
-                    raise UserError(_("{}: {}").format(wo.name, str(e)))
-                except AttributeError:
-                    # If enterprise module mrp_workorder module
-                    # is not installed quality check is not necessary
-                    pass
-
-    def button_pending(self):
-        for nl in self:
-            is_user_working = nl.workorder_id.is_user_working
-            if nl.working_state != "blocked" and is_user_working and \
-                    nl.state not in ("done", "pending", "ready", "cancel"):
-                nl.workorder_id.with_context(
-                    from_nest=True).button_pending()
-
-    def button_scrap(self):
-        for nest_line in self:
-            if nest_line.state not in ("confirmed", "cancel"):
-                nest_line.workorder_id.with_context(
-                    from_nest=True).button_scrap()
-
-    def button_quality_alert(self):
-        return self.workorder_id.button_quality_alert()
-
-    def button_unblock(self):
-        return self.workcenter_id.unblock()
-
-    def button_change_ok_line(self):
-        self.ok_line = not self.ok_line
-        return self.nest_line_form_view(self.id)
-
-    def open_line(self):
-        self.ensure_one()
-        if self.nest_id.nested_line_ids:
-            view_ref = self.env["ir.model.data"].get_object_reference(
-                "mrp_workorder_grouping_by_material",
-                "mrp_workorder_nest_line_form")
-            view_id = view_ref and view_ref[1] or False,
-
-        return {
-            "name": "Workorder nest line",
-            "domain": [],
-            "res_model": "mrp.workorder.nest.line",
-            "res_id": self.id,
-            "type": "ir.actions.act_window",
-            "view_mode": "form",
-            "view_type": "form",
-            "view_id": view_id,
-            "context": {},
-            "target": "new",
-        }
