+++ mrp_workorder_grouping_by_material/views/mrp_workorder_nest_line_view.xml
@@ -1,0 +1,100 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<odoo>
+    <record id="mrp_workorder_nest_line_view_form" model="ir.ui.view">
+        <field name="model">mrp.workorder.nest.line</field>
+        <field name="arch" type="xml">
+            <form>
+                <header>
+<!--                                    <button name="button_finish" type="object" string="Finish Order"-->
+<!--                                            attrs="{'invisible': ['|', ('state', '!=', 'progress'), ('is_produced', '=', False)]}"-->
+<!--                                            class="btn-info"/>-->
+<!--                                    <button name="button_start" type="object" string="Start Working"-->
+<!--                                            attrs="{'invisible': ['|', ('working_state', '=', 'blocked'), ('state', '!=', 'pending')]}"/>-->
+<!--                                    <button name="button_start" type="object" string="Start Working"-->
+<!--                                            attrs="{'invisible': ['|', ('working_state', '=', 'blocked'), ('state', '!=', 'ready')]}"-->
+<!--                                            class="btn-success"/>-->
+<!--                                    <button name="record_production" type="object" string="Done" class="btn-success"-->
+<!--                                            attrs="{'invisible': ['|', '|', '|', ('is_produced', '=', True), ('working_state', '=', 'blocked'), ('state', '!=', 'progress'), ('is_user_working', '=', False)]}"/>-->
+<!--                                    <button name="button_pending" type="object" string="Pause" class="btn-warning"-->
+<!--                                            attrs="{'invisible': ['|', '|', ('working_state', '=', 'blocked'), ('state', 'in', ('done', 'pending', 'ready', 'cancel')), ('is_user_working', '=', False)]}"/>-->
+<!--                                    <button name="button_start" type="object" string="Continue Production"-->
+<!--                                            attrs="{'invisible': ['|', '|', '|', ('production_state', '=', 'done'), ('working_state', '=', 'blocked'), ('is_user_working', '=', True), ('state', '!=', 'progress')]}"/>-->
+<!--                                    <button name="button_scrap" type="object" string="Scrap"-->
+<!--                                            attrs="{'invisible': [('state', 'in', ('done', 'cancel'))]}"/>-->
+                    <button name="mrp.act_mrp_block_workcenter_wo" type="action"
+                            context="{'default_workcenter_id': workcenter_id}" string="Block"
+                            class="btn-danger"
+                            attrs="{'invisible': ['|', '|', ('working_state', '=', 'blocked'), ('state', 'in', ('done', 'pending', 'ready', 'cancel')), ('is_user_working', '=', False)]}"/>
+                    <button name="button_unblock" type="object"
+                            string="Unblock" class="btn-danger"
+                            attrs="{'invisible': [('working_state', '!=', 'blocked')]}"/>
+                    <button name="button_quality_alert" type="object" string="Quality Alert"
+                            attrs="{'invisible': [('state', 'in', ('done', 'cancel'))]}"/>
+                    <field name="state" widget="statusbar"
+                           statusbar_visible="pending,ready,progress,done"/>
+                </header>
+                <sheet>
+                    <notebook>
+                        <page name="data" string="Data">
+                    <group>
+                        <group>
+                            <field name="nest_id" invisible="1"/>
+                            <field name="possible_workorder_ids"
+                                   string="Possible Workorders"
+                                   invisible="1"/>
+                            <field name="workorder_id"
+                                   options="{'no_create': True}"
+                                   domain="[('id', 'in', possible_workorder_ids)]"/>
+                            <field name="name" readonly="1"/>
+                            <field name="product_id" readonly="1"/>
+                            <field name="qty_producing"/>
+                            <field name="company_id" invisible="1"/>
+                            <field name="finished_lot_id"
+                                   string="Product to Produce Lot"
+                                   context="{'default_product_id': product_id, 'default_company_id': company_id}"/>
+                            <field name="qty_produced" readonly="1"/>
+                            <field name="qty_remaining" readonly="1"/>
+                            <field name="qty_production" readonly="1"/>
+                        </group>
+                        <group>
+                            <field name="product_uom_id" readonly="1"/>
+                            <field name="workcenter_id" readonly="1"/>
+                            <field name="production_id" readonly="1"/>
+                            <field name="date_planned_start" readonly="1"/>
+                            <field name="working_state" invisible="1"/>
+                            <field name="production_state" invisible="1"/>
+                            <field name="is_user_working" invisible="1"/>
+                            <field name="is_produced" invisible="1"/>
+                            <field name="ok_line" nolabel="1" readonly="1"/>
+                            <button name="button_change_ok_line" type="object" string="OK line"/>
+                        </group>
+                    </group>
+                        </page>
+                        <page name="worksheet" string="Work Instruction"
+                              attrs="{'invisible': [('worksheet', '=', False)]}">
+                            <field name="worksheet" widget="pdf_viewer"/>
+                        </page>
+                    </notebook>
+                </sheet>
+            </form>
+        </field>
+    </record>
+
+    <record model="ir.ui.view"
+            id="mrp_workorder_nest_line_wizard_form">
+        <field name="model">mrp.workorder.nest.line</field>
+        <field name="inherit_id" ref="mrp_workorder_nest_line_view_form" />
+        <field name="mode">primary</field>
+        <field name="arch" type="xml">
+            <field name="state" position="before">
+                <button name="button_get_previous_line"
+                        type="object"
+                        icon="fa-chevron-left"/>
+                <button name="button_get_next_line"
+                        type="object"
+                        icon="fa-chevron-right"/>
+            </field>
+        </field>
+    </record>
+
+</odoo>
