--- mrp_workorder_grouping_by_material/wizard/nested_new_line_action_menu.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<odoo>
-
-
-</odoo>
