+++ mrp_workorder_grouping_by_material/wizard/nested_new_line_view.xml
@@ -12,7 +12,9 @@
                 </group>
                 <field name="line_ids">
                     <tree editable="top">
-                        <field name="workorder_id" options="{'no_create': True}"/>
+                        <field name="workorder_id"
+                               domain="[('state', 'not in', ['done', 'cancel'])]"
+                               options="{'no_create': True}"/>
                         <field name="product_id" options="{'no_create': True}"/>
                         <field name="qty_producing"/>
                         <field name="qty_produced" readonly="1"/>
@@ -32,7 +34,6 @@
     <record id="nested_new_line_wizard_action" model="ir.actions.act_window">
         <field name="name">Add to Nest</field>
         <field name="res_model">nested.new.line</field>
-<!--        <field name="binding_model">mrp.workorder</field>-->
         <field name="binding_model_id" ref="mrp.model_mrp_workorder" />
         <field name="binding_view_types">list</field>
         <field name="view_mode">form</field>
